downgraded to SI. Made the timestamps to uint64 isntead of ints

+++ snapflow/txn/mvcc_storage.h
@@ -8,7 +8,7 @@
 
 
 struct TimeStamp {
-  int timestamp;
+  uint64 timestamp;
   Txn* txn;
   bool edit_bit;
 };
